requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
pydantic>=2.8.2
openai>=1.40.0
